*** empty log message ***
git-svn-id: http://svn.vuze.com/public/client@7084 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/com/aelitis/azureus/plugins/startstoprules/defaultplugin/StartStopRulesDefaultPlugin.java
@@ -1701,8 +1701,9 @@ public class StartStopRulesDefaultPlugin
 	            else { // Peers <= Seeds
 	              newSR += num_peers_excluding_us * 1000 / num_seeds_excluding_us;
 	            }*/
-	            float x =  num_seeds_excluding_us  / num_peers_excluding_us;
+	            float x =  (float)(num_seeds_excluding_us)  / num_peers_excluding_us;
 	            newSR += 10000 / ((x+1)*(x+1));
+	            //System.out.println(dl.getName() + ": Seeds: " + num_seeds_excluding_us + " - Peers: " + num_peers_excluding_us + " - Ratio: " + x + " Rank: " + newSR);
 	          }
 	
 	          if (bPreferLargerSwarms)
